requirements.txt
anyio==3.5.0
apturl==0.5.2
asgiref==3.5.1
astroid==2.11.7
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
blinker==1.4
Brlapi==0.8.3
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
command-not-found==0.3
comtypes==1.1.11
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.0
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
dill==0.3.5.1
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
dj-database-url==0.5.0
Django==4.0.4
django-allauth==0.50.0
django-bootstrap4==22.1
django-countries==7.3.2
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-registration==3.3
django-tinymce==3.5.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.20.0
duplicity==0.8.21
fastapi==0.77.1
fasteners==0.14.1
filelock==3.6.0
future==0.18.2
gunicorn==20.1.0
gyp==0.1
h11==0.13.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==3.2.0
keyring==23.5.0
language-selector==0.1
lark==1.1.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.7.1
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.1
mccabe==0.7.0
monotonic==1.6
more-itertools==8.10.0
MouseInfo==0.1.3
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
outcome==1.1.0
packaging==21.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pipenv==2022.5.2
platformdirs==2.5.2
protobuf==3.12.4
psutil==5.9.1
psycopg2==2.9.3
ptyprocess==0.7.0
PyAutoGUI==0.9.53
pycairo==1.20.1
pycodestyle==2.8.0
pycparser==2.21
pycups==2.0.1
pydantic==1.9.0
PyGetWindow==0.0.9
Pygments==2.11.2
PyGObject==3.42.1
pyjokes==0.6.0
PyJWT==2.3.0
pylint==2.14.5
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyRFC3339==1.1
pyrsistent==0.18.1
PyScreeze==0.1.28
PySocks==1.7.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-decouple==3.6
python-dotenv==0.20.0
python3-openid==3.2.0
python3-xlib==0.15
pyttsx3==2.90
pytweening==1.0.4
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
SecretStorage==3.3.1
selenium==4.1.5
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
sqlparse==0.4.2
starlette==0.19.1
systemd-python==234
texttable==1.6.4
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.2.0
ubuntu-advantage-tools==27.11.2
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.9
usb-creator==0.3.7
virtualenv==20.14.1
virtualenv-clone==0.5.7
wadllib==1.3.6
websocket-client==0.59.0
whitenoise==6.1.0
wikipedia==1.4.0
wrapt==1.14.1
wsproto==1.1.0
xdg==5
xkit==0.0.0
zipp==1.0.0
